requirements.txt
watchdog # >=3.0
pillow # >=10.3
pytesseract # >=0.3
# pdf2image # >=1.16 # Replaced by PyMuPDF
# poppler-utils # System dependency for pdf2image, no longer needed
# pyzbar # >=0.1 # Replaced by opencv-python-headless
PyMuPDF # For PDF processing
opencv-python-headless # For QR code detection and image manipulation
openai # >=1.22
pydantic # >=2.7
pydantic-settings # >=2.2
ruamel.yaml # >=0.18
markdown-it-py # >=3.0
python-dotenv  # >=1.0
fastapi # >=0.100
uvicorn[standard] # >=0.20
Jinja2>=3.0 # For facsimile templating
